+++ ToDoList/Sources/Constants.swift
@@ -1,0 +1,12 @@
+//
+//  Constants.swift
+//  ToDoList
+//
+//  Created by Sergey Vysotsky on 01.03.2022.
+//
+
+import Foundation
+
+enum Constants {
+    static let taskCellIdentifier = "IdCell"
+}
